manual test of collapsable section
Testing the collapsible section for Equivalent XPath for the term
Audience. If this works, the changes need to be adjusted in the build
script.

Please note: it is important that the line of the XPaths as well as the
following two closing tags: </details> and </td> are NOT indented. Also
there needs to be an empty line after the summary tag.